+++ bot/stickfix.py
@@ -7,9 +7,11 @@
 """
 from typing import List
 
-from telegram.ext import CommandHandler, Dispatcher, Updater
+from telegram import Update
+from telegram.ext import CallbackContext, CommandHandler, Dispatcher, Updater
 
-import bot.messages
+from bot.database.storage import StickfixDB
+from bot.database.users import StickfixUser
 from bot.logger import StickfixLogger
 
 
@@ -22,6 +24,7 @@ class Stickfix:
     __dispatcher: Dispatcher
     __logger: StickfixLogger
     __admins: List[int]
+    __user_db: StickfixDB
 
     def __init__(self, token: str, admins: List[int]):
         """ Initializes the bot.
@@ -36,6 +39,7 @@ class Stickfix:
         self.__admins = admins
         self.__dispatcher = self.__updater.dispatcher
         self.__setup_handlers()
+        self.__user_db = StickfixDB("stickfix-user-DB")
 
     def run(self) -> None:
         """ Runs the bot.   """
@@ -48,4 +52,17 @@ class Stickfix:
         self.__updater = Updater(token, use_context=True)
 
     def __setup_handlers(self):
-        self.__dispatcher.add_handler(CommandHandler("start", bot.messages.send_hello_message))
+        self.__dispatcher.add_handler(CommandHandler("start", self.__send_hello_message))
+
+    def __send_hello_message(self, update: Update, context: CallbackContext) -> None:
+        """ Answers the /start command with a hello sticker and adds the user to the database. """
+        chat_id = update.effective_chat.id
+        context.bot.send_sticker(chat_id, sticker='CAADBAADTAADqAABTgXzVqN6dJUIXwI')
+        if chat_id not in self.__user_db:
+            self.__create_user(chat_id)
+            self.__logger.info(f"User {chat_id} was added to the database.")
+
+    def __create_user(self, chat_id: str) -> None:
+        """ Creates and adds a user to the database.    """
+        user = StickfixUser(chat_id)
+        self.__user_db.add_item(chat_id, user)