+++ Changelog.md
@@ -2,5 +2,6 @@
 
 ## Version 2.2
 
+- [B.3] Modified database to use Yaml
 - [B.2] Updated ``/start`` command to new API version
 - [B.1] Added stickfix logger